requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pandas>=2.2.0
streamlit==1.28.1